Fixed: tweets grammar and file structure.
