+++ client/src/components/CheckList/CheckList.tsx
@@ -41,7 +41,13 @@ export function CheckList({
                 onChange={() => handleCheckBoxChange(id, text)}
                 isChecked={isChecked}
               >
-                <Text pos="relative" top="-2px" color="black.200" overflowWrap="break-word" wordBreak="break-word">
+                <Text
+                  pos="relative"
+                  top="-2px"
+                  color="black.200"
+                  overflowWrap="break-word"
+                  wordBreak="break-word"
+                >
                   {text}
                 </Text>
               </Checkbox>
