--- client/src/utils/differentTweets.ts
@@ -1,23 +1,0 @@
-export const Tweets = [
-    'I%20am%20super%20excited%20to%20announce%20that%20I%20have%20completed%20levelZero%20of%20%40neogcamp%20and%20have%20successfully%20submitted%20my%20mark15%20portfolio.%0AWish%20me%20luck%20for%20admissions.%20%0A%23neogcamp%20%23webdevelopment%20%23neog2022',
-    'Can%27t%20believe%20how%20far%20I%20have%20come%20in%20the%20last%20few%20months%20in%20my%20web%20development%20journey.%20Mark15%20portfolio%20successfully%20submitted.%0AOff%20to%20prepare%20for%20interviews.%0AWish%20me%20luck.%0A%23neog2022%20%23webdevelopment%20%23neogcamp%20',
-    'OMG%21%20I%20have%20done%2015%20projects%20in%20just%20levelZero.%20Can%27t%20wait%20to%20see%20what%20there%20will%20be%20in%20levelOne.%0AOff%20to%20prepare%20for%20interviews.%0AWish%20me%20luck.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Done%20and%20dusted%20with%20levelZero%20and%20my%20web%20dev%20journey%20starts%20with%20a%20bang.%20Not%20just%20JavaScript%20but%20React%2C%20VS%20Code%2C%20and%20made%20so%20many%20projects.%0AExcited%20to%20prepare%20for%20levelOne%20interviews.%20%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'The%20fact%20that%20I%20have%20made%2010%2B%20projects%20and%20published%202%20blogs%20on%20my%20own%20proves%20how%20much%20I%20enjoy%20learning%20web%20development.%0AMark15%20portfolio%20successfully%20submitted.%20Off%20to%20prepare%20for%20interviews.%0AWish%20me%20luck.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Finally%2C%20I%20can%20tell%20my%20friends%20I%20have%20my%20own%20website.%0ANot%20just%20a%20website%20but%20web%20apps%20with%20blogs%20as%20well.%0ASubmitted%20my%20mark15%20portfolio.%20Getting%20ready%20for%20the%20interviews.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Basics%20done.%0AReady%20to%20get%20advanced%20with%20%40neogcamp%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'React%20%3A%20%E2%98%91%EF%B8%8F%0AGit%20%3A%20%E2%98%91%EF%B8%8F%0AVS%20Code%20%3A%20%E2%98%91%EF%B8%8F%0AVanillaJS%20%3A%20%E2%98%91%EF%B8%8F%0ABasics%20done.%0AReady%20to%20become%20advanced%20in%20web%20development%20%40neogcamp.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Done%20and%20dusted%20with%20basic%20and%20level%200.%0AMERN%20stack%20here%20I%20come.%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'A%20good%20teacher%20can%20make%20so%20much%20difference%20in%20life.%0AIf%20anybody%20disagrees%2C%20please%20see%20my%20projects%20and%20my%20portfolio.%0ADone%20with%20levelZero.%0AGetting%20ready%20for%20levelOne%20admissions.%0AWish%20me%20luck.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'It%27s%20unbelievable%21%20Mark15%20is%20done.%20Here%20I%20come%2C%20Neog%2722.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'This%20marks%20the%20completion%20of%20my%20mark15%20portfolio.%20I%20have%20developed%208%20web%20apps%20including%20my%20portfolio%20and%20written%20a%20few%20blogs.%20Looking%20back%2C%20I%20see%20that%20my%20learning%20has%20grown%20exponentially%21%20Thankyou%20%40%20neogcamp.%20Gearing%20up%20for%20interviews.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'I%20wouldn%27t%20miss%20the%20chance%20to%20learn%20from%20%40tanaypratap.%20I%20have%20completed%20my%20Mark15%20portfolio.%20See%20you%20inside%20the%20camp%2C%20everyone%21%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Now%20that%20I%27ve%20completed%20levelZero%20and%20submitted%20my%20portfolio%2C%20I%20cannot%20wait%20to%20move%20up%20to%20levelOne.%20%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'I%20cannot%20wait%20to%20see%20what%20levelOne%20has%20in%20store%20for%20me%20if%20levelZero%20is%20this%20amazing.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'I%27m%20leveling%20up%20with%20%40neogcamp%20and%20Tanay%20Pratap.%20Finally%2C%20LevelZero%20is%20done%20and%20dusted.%20Gearing%20up%20for%20the%20interviews.%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Level%200%C2%A0%20%E2%9C%85%0AMark15%20checklist%20%E2%9C%85%0APortfolio%20submission%20%E2%9C%85%0ATo-be-next%0APrepare%20for%20the%20interview.%0A%0AI%20am%20very%20excited%20for%20Neog%E2%80%9922.%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    '%40neogcamp%27s%20Mark15%20portfolio%20submission%20is%20done%20%26%20dusted.%20I%20am%20so%20excited%20to%20sit%20for%20the%20interviews%20and%20be%20accepted%20to%20levelOne.%20%0ANeogrammers%2C%20Send%20a%20%F0%9F%9A%80%20in%20the%20comments%20if%20you%20have%20applied%20for%20neog%2722%20as%20well.%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'Web%20apps%3A%20%E2%98%91%EF%B8%8F%0A%0APersonal%20Portfolio%20%3A%20%E2%98%91%EF%B8%8F%0A%0ABlogs%20%3A%20%E2%98%91%EF%B8%8F%0A%0AYayyyy%21%20Mark15%20is%20Done.%20Let%27s%20get%20ready%20for%20the%20interviews.%20%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    '-%20HTML%0A-%20CSS%0A-%20git%0A-%20Github%0A-%20VanillaJS%0A-%20React%0A%0ALearned%20all%20of%20these%20in%20levelZero.%20Submitted%20by%20Mark15%20portfolio.%20%F0%9F%9A%80%0A%0AExcited%20to%20see%20what%20%40neogcamp%20camp%20has%20in%20store%20for%20us%20in%20levelOne.%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp',
-    'How%20is%20the%20josh%3A%20HIGH%20sir%21%21%0ASubmitted%20my%20mark15%20portfolio%20and%20can%20not%20be%20more%20excited.%20Wish%20me%20luck%20for%20interviews.%20%0A%0A%23neog2022%20%23webdevelopment%20%23neogcamp'
-  ]
